fix: readme path not correctly expanded
When not specifiying a readme config property, the readmy path is set to the directory of the lerna package
Before, it was only checked if the readme is not none but as lerna sets the property to `''` when no config is set, the check becomes truthy and the plugin tries to read the directory.
Added a check if `this.readme` is truthy to solve this issue, as `''` is falsy.